test(chat): split ACP cohort helpers
